clean up client-duckdb types
