+++ packages/universal-sql/src/client-duckdb/both.d.ts
@@ -1,0 +1,6 @@
+/**
+ * Converts an Apache Arrow table to a Javascript array.
+ * @param {import("apache-arrow").Table} table
+ * @returns {any[]}
+ */
+export function arrowTableToJSON(table: import('apache-arrow').Table): any[];
