+++ packages/universal-sql/src/client-duckdb/both.js
@@ -1,0 +1,44 @@
+import { Type } from 'apache-arrow';
+
+/**
+ * Converts an Apache Arrow type to an Evidence type.
+ *
+ * @param {import("apache-arrow").Type} type
+ */
+function apacheToEvidenceType(type) {
+	switch (
+		type.typeId // maybe just replace with `typeof`
+	) {
+		case Type.Date:
+			return 'date';
+		case Type.Float:
+		case Type.Int:
+			return 'number';
+		case Type.Bool:
+			return 'boolean';
+		case Type.Dictionary:
+		default:
+			return 'string';
+	}
+}
+
+/**
+ * Converts an Apache Arrow table to a Javascript array.
+ * @param {import("apache-arrow").Table} table
+ * @returns {any[]}
+ */
+export function arrowTableToJSON(table) {
+	if (table == null) return [];
+	const arr = table.toArray();
+
+	Object.defineProperty(arr, '_evidenceColumnTypes', {
+		enumerable: false,
+		value: table.schema.fields.map((field) => ({
+			name: field.name,
+			evidenceType: apacheToEvidenceType(field.type),
+			typeFidelity: 'precise'
+		}))
+	});
+
+	return arr;
+}
