+++ packages/universal-sql/src/client-duckdb/browser.d.ts
@@ -1,4 +1,30 @@
+/**
+ * Initializes the database.
+ *
+ * @returns {Promise<void>}
+ */
 export function initDB(): Promise<void>;
-export function query(query: string): Promise<import('apache-arrow').Table | null>;
-export function setParquetURL(url: string): Promise<void>;
+
+/**
+ * Queries the database with the given SQL statement.
+ *
+ * @param {string} sql
+ * @returns {Promise<ReturnType<import("@duckdb/duckdb-wasm").AsyncDuckDBConnection['query']> | null>}
+ */
+export function query(sql: string): Promise<import('apache-arrow').Table | null>;
+
+/**
+ * Adds a new view to the database, pointing to the provided parquet URL.
+ *
+ * @param {string} table
+ * @param {string} url
+ * @returns {Promise<void>}
+ */
+export function setParquetURL(table: string, url: string): Promise<void>;
+
+/**
+ * Converts an Apache Arrow table to a Javascript array.
+ * @param {import("apache-arrow").Table} table
+ * @returns {any[]}
+ */
 export function arrowTableToJSON(table: import('apache-arrow').Table): any[];
